pyproject.toml
[project]
name = "teaorcoffee"
version = "0.1.0"
description = ""
authors = [
    {name = "vaibhavsinghtaurus",email = "vaibhav.singh@thinktaurus.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.6,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "teaorcoffee", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
